third_party/blink/web_tests/platform/generic/fast/css/vertical-align-length-copy-bug-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=71996

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.verticalAlign is "-1px"
PASS test.offsetTop is 10
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
